requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
argon2-cffi>=21.3.0
cryptography>=41.0.0
redis>=4.6.0
PyJWT>=2.8.0



